decklink: Fix various compiler warnings with gcc

+++ sys/decklink/gstdecklinkaudiosrc.cpp
@@ -405,12 +405,12 @@ gst_decklink_audio_src_create (GstPushSrc * bsrc, GstBuffer ** buffer)
 
   // Convert to the sample numbers
   end_offset = gst_util_uint64_scale (end_time, self->info.rate, GST_SECOND);
-  if (end_offset >= sample_count)
+  if (end_offset >= (guint64) sample_count)
     start_offset = end_offset - sample_count;
   else
     start_offset = 0;
 
-  if (self->next_offset == -1) {
+  if (self->next_offset == (guint64) -1) {
     discont = TRUE;
   } else {
     guint64 diff, max_sample_diff;
@@ -445,7 +445,7 @@ gst_decklink_audio_src_create (GstPushSrc * bsrc, GstBuffer ** buffer)
 
   if (discont) {
     // Have discont, need resync and use the capture timestamps
-    if (self->next_offset != -1)
+    if (self->next_offset != (guint64) -1)
       GST_INFO_OBJECT (self, "Have discont. Expected %"
           G_GUINT64_FORMAT ", got %" G_GUINT64_FORMAT,
           self->next_offset, start_offset);
